moved conversation and adapter into separate files
