+++ backend/src/kg/db/queries.py
@@ -1,4 +1,5 @@
 from .util import run_query
+from neomodel import db
 
 def get_all_papers(kg):
     cypher = """
@@ -46,3 +47,56 @@ def search_papers_by_title(kg, searchTerm):
     """
     result = run_query(kg, cypher, params={'searchTerm': searchTerm})
     return result
+
+def retrieve_similar_chunks(embedding, k=30):
+    query = """
+        MATCH (c:Chunk)
+        WITH DISTINCT c, vector.similarity.cosine(c.textEmbedding, $embedding) AS score
+        WHERE score > 0.5 
+        ORDER BY score DESC LIMIT $k
+        RETURN c.text AS text, score, c.source AS source, c.chunkId AS chunkId
+    """
+    
+    results, meta = db.cypher_query(
+        query, 
+        {'embedding': embedding, 'k': k}
+    )
+    
+    return [
+        {
+            'text': row[0],
+            'score': row[1],
+            'metadata': {
+                'source': row[2],
+                'chunkId': row[3]
+            }
+        }
+        for row in results
+    ]
+
+def retrieve_similar_abstracts(embedding, k=30):
+    query = """
+        MATCH (p:Paper)
+        WHERE p.abstract IS NOT NULL AND p.title IS NOT NULL
+        WITH DISTINCT p, vector.similarity.cosine(p.abstract_embedding, $embedding) AS score
+        WHERE score > 0.5 
+        ORDER BY score DESC LIMIT $k
+        RETURN 'Title: ' + p.title + '\\n\\n' + 'Abstract: ' + p.abstract AS text, 
+               score, p.paper_id AS paper_id
+    """
+    
+    results, meta = db.cypher_query(
+        query, 
+        {'embedding': embedding, 'k': k}
+    )
+    
+    return [
+        {
+            'text': row[0],
+            'score': row[1],
+            'metadata': {
+                'paper_id': row[2]
+            }
+        }
+        for row in results
+    ]
